// src/generated/types/account.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { AccountDescriptor } from './accountDescriptor';
import type { AccountDevicesStatus } from './accountDevicesStatus';

/**
 * Account data
 */
export interface Account {
  id?: number;
  name?: string;
  descriptor?: AccountDescriptor;
  uid?: string;
  currency?: string;
  devicesStatus?: AccountDevicesStatus;
}