// src/generated/types/dnetSiteMappingsDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */

/**
 * Dnet site mappings data
 */
export interface DnetSiteMappingsDto {
  id?: number;
  /** Unique alphanumeric UID of this site */
  uid?: string;
  /** Unique alphanumeric UID of the account to which this site belongs */
  accountUid?: string;
  name?: string;
  description?: string;
  dattoNetworkingNetworkIds?: number[];
  portalUrl?: string;
}